update tests to avoid duplicated
