+++ src/__tests__/useTreatments.test.tsx
@@ -1,157 +1,24 @@
-import React from 'react';
-import { render, act } from '@testing-library/react';
-
 /** Mocks */
-import { mockSdk, Event } from './testUtils/mockSplitSdk';
-jest.mock('@splitsoftware/splitio/client', () => {
-  return { SplitFactory: mockSdk() };
-});
-import { SplitFactory as SplitSdk } from '@splitsoftware/splitio/client';
-import { sdkBrowser } from './testUtils/sdkConfigs';
-jest.mock('../constants', () => {
-  const actual = jest.requireActual('../constants');
-  return {
-    ...actual,
-    getControlTreatmentsWithConfig: jest.fn(actual.getControlTreatmentsWithConfig),
-  };
-});
-import { CONTROL_WITH_CONFIG, getControlTreatmentsWithConfig } from '../constants';
-const logSpy = jest.spyOn(console, 'log');
+const useSplitTreatmentsMock = jest.fn();
+jest.mock('../useSplitTreatments', () => ({
+  useSplitTreatments: useSplitTreatmentsMock
+}));
 
 /** Test target */
-import { SplitFactory } from '../SplitFactory';
-import { SplitClient } from '../SplitClient';
 import { useTreatments } from '../useTreatments';
 
 describe('useTreatments', () => {
 
-  const featureFlagNames = ['split1'];
-  const attributes = { att1: 'att1' };
+  test('calls useSplitTreatments with the correct arguments and returns the treatments.', () => {
+    const names = ['someFeature'];
+    const attributes = { someAttribute: 'someValue' };
+    const treatments = { someFeature: { treatment: 'on', config: null } };
+    useSplitTreatmentsMock.mockReturnValue({ treatments, isReady: false });
 
-  test('returns the treatments evaluated by the client at Split context updated by SplitFactory, or control if the client is not operational.', () => {
-    const outerFactory = SplitSdk(sdkBrowser);
-    const client: any = outerFactory.client();
-    let treatments: SplitIO.TreatmentsWithConfig;
+    expect(useTreatments(names, attributes, 'someKey')).toBe(treatments);
 
-    render(
-      <SplitFactory factory={outerFactory} >
-        {React.createElement(() => {
-          treatments = useTreatments(featureFlagNames, attributes);
-          return null;
-        })}
-      </SplitFactory>
-    );
-
-    // returns control treatment if not operational (SDK not ready or destroyed), without calling `getTreatmentsWithConfig` method
-    expect(client.getTreatmentsWithConfig).not.toBeCalled();
-    expect(treatments!).toEqual({ split1: CONTROL_WITH_CONFIG });
-
-    // once operational (SDK_READY), it evaluates feature flags
-    act(() => client.__emitter__.emit(Event.SDK_READY));
-
-    expect(client.getTreatmentsWithConfig).toBeCalledWith(featureFlagNames, attributes);
-    expect(client.getTreatmentsWithConfig).toHaveReturnedWith(treatments);
-  });
-
-  test('returns the Treatments from the client at Split context updated by SplitClient, or control if the client is not operational.', async () => {
-    const outerFactory = SplitSdk(sdkBrowser);
-    const client: any = outerFactory.client('user2');
-    let treatments: SplitIO.TreatmentsWithConfig;
-
-    render(
-      <SplitFactory factory={outerFactory} >
-        <SplitClient splitKey='user2' >
-          {React.createElement(() => {
-            treatments = useTreatments(featureFlagNames, attributes);
-            return null;
-          })}
-        </SplitClient>
-      </SplitFactory>
-    );
-
-    // returns control treatment if not operational (SDK not ready or destroyed), without calling `getTreatmentsWithConfig` method
-    expect(client.getTreatmentsWithConfig).not.toBeCalled();
-    expect(treatments!).toEqual({ split1: CONTROL_WITH_CONFIG });
-
-    // once operational (SDK_READY_FROM_CACHE), it evaluates feature flags
-    act(() => client.__emitter__.emit(Event.SDK_READY_FROM_CACHE));
-
-    expect(client.getTreatmentsWithConfig).toBeCalledWith(featureFlagNames, attributes);
-    expect(client.getTreatmentsWithConfig).toHaveReturnedWith(treatments);
-  });
-
-  test('returns the Treatments from a new client given a splitKey, and re-evaluates on SDK events.', () => {
-    const outerFactory = SplitSdk(sdkBrowser);
-    const client: any = outerFactory.client('user2');
-    let renderTimes = 0;
-
-    render(
-      <SplitFactory factory={outerFactory} >
-        {React.createElement(() => {
-          const treatments = useTreatments(featureFlagNames, attributes, 'user2');
-
-          renderTimes++;
-          switch (renderTimes) {
-            case 1:
-              // returns control if not operational (SDK not ready), without calling `getTreatmentsWithConfig` method
-              expect(client.getTreatmentsWithConfig).not.toBeCalled();
-              expect(treatments).toEqual({ split1: CONTROL_WITH_CONFIG });
-              break;
-            case 2:
-            case 3:
-              // once operational (SDK_READY or SDK_READY_FROM_CACHE), it evaluates feature flags
-              expect(client.getTreatmentsWithConfig).toHaveBeenLastCalledWith(featureFlagNames, attributes);
-              expect(client.getTreatmentsWithConfig).toHaveLastReturnedWith(treatments);
-              break;
-            default:
-              throw new Error('Unexpected render');
-          }
-
-          return null;
-        })}
-      </SplitFactory>
-    );
-
-    act(() => client.__emitter__.emit(Event.SDK_READY_FROM_CACHE));
-    act(() => client.__emitter__.emit(Event.SDK_READY));
-    act(() => client.__emitter__.emit(Event.SDK_UPDATE)); // should not trigger a re-render by default
-    expect(client.getTreatmentsWithConfig).toBeCalledTimes(2);
-  });
-
-  // THE FOLLOWING TEST WILL PROBABLE BE CHANGED BY 'return a null value or throw an error if it is not inside an SplitProvider'
-  test('returns Control Treatments if invoked outside Split context.', () => {
-    let treatments;
-
-    render(
-      React.createElement(() => {
-        treatments = useTreatments(featureFlagNames, attributes);
-        return null;
-      })
-    );
-    expect(getControlTreatmentsWithConfig).toBeCalledWith(featureFlagNames);
-    expect(getControlTreatmentsWithConfig).toHaveReturnedWith(treatments);
-  });
-
-  /**
-   * Input validation. Passing invalid feature flag names or attributes while the Sdk
-   * is not ready doesn't emit errors, and logs meaningful messages instead.
-   */
-  test('Input validation: invalid "names" and "attributes" params in useTreatments.', (done) => {
-    render(
-      React.createElement(() => {
-        // @ts-expect-error Test error handling
-        let treatments = useTreatments('split1');
-        expect(treatments).toEqual({});
-        // @ts-expect-error Test error handling
-        treatments = useTreatments([true]);
-        expect(treatments).toEqual({});
-
-        done();
-        return null;
-      })
-    );
-    expect(logSpy).toBeCalledWith('[ERROR] split names must be a non-empty array.');
-    expect(logSpy).toBeCalledWith('[ERROR] you passed an invalid split name, split name must be a non-empty string.');
+    expect(useSplitTreatmentsMock).toHaveBeenCalledTimes(1);
+    expect(useSplitTreatmentsMock).toHaveBeenCalledWith({ names, attributes, splitKey: 'someKey' });
   });
 
 });
